libublk: move run_target() into UblkCtrl
Prepare for replacing UblkSession with UblkCtrlBuilder
