+++ src/ctrl.rs
@@ -5,7 +5,7 @@ use io_uring::{cqueue, opcode, squeue, types, IoUring};
 use log::{error, trace};
 use serde::Deserialize;
 use std::os::unix::io::{AsRawFd, RawFd};
-use std::sync::RwLock;
+use std::sync::{Arc, RwLock};
 use std::{
     fs,
     io::{Read, Write},
@@ -1207,6 +1207,102 @@ impl UblkCtrl {
         if Path::new(&ctrl.run_path()).exists() {
             fs::remove_file(ctrl.run_path()).map_err(UblkError::OtherIOError)?;
         }
+        Ok(0)
+    }
+
+    fn create_queue_handlers<Q>(
+        &self,
+        dev: &Arc<UblkDev>,
+        q_fn: Q,
+    ) -> Vec<std::thread::JoinHandle<()>>
+    where
+        Q: FnOnce(u16, &UblkDev) + Send + Sync + Clone + 'static,
+    {
+        use std::sync::mpsc;
+
+        let mut q_threads = Vec::new();
+        let nr_queues = dev.dev_info.nr_hw_queues;
+
+        let (tx, rx) = mpsc::channel();
+
+        for q in 0..nr_queues {
+            let _dev = Arc::clone(dev);
+            let _tx = tx.clone();
+
+            let mut affinity = UblkQueueAffinity::new();
+            self.get_queue_affinity(q as u32, &mut affinity).unwrap();
+            let mut _q_fn = q_fn.clone();
+
+            q_threads.push(std::thread::spawn(move || {
+                //setup pthread affinity first, so that any allocation may
+                //be affine to cpu/memory
+                unsafe {
+                    libc::pthread_setaffinity_np(
+                        libc::pthread_self(),
+                        affinity.buf_len(),
+                        affinity.addr() as *const libc::cpu_set_t,
+                    );
+                }
+                _tx.send((q, unsafe { libc::gettid() })).unwrap();
+
+                unsafe {
+                    const PR_SET_IO_FLUSHER: i32 = 57; //include/uapi/linux/prctl.h
+                    libc::prctl(PR_SET_IO_FLUSHER, 0, 0, 0, 0);
+                };
+
+                _q_fn(q, &_dev);
+            }));
+        }
+
+        for _q in 0..nr_queues {
+            let (qid, tid) = rx.recv().unwrap();
+            if self.configure_queue(dev, qid, tid).is_err() {
+                println!(
+                    "create_queue_handler: configure queue failed for {}-{}",
+                    dev.dev_info.dev_id, qid
+                );
+            }
+        }
+
+        q_threads
+    }
+
+    /// Run ublk daemon and kick off the ublk device, and `/dev/ublkbN` will be
+    /// created and visible to userspace.
+    ///
+    /// # Arguments:
+    ///
+    /// * `tgt_fn`: target initialization handler
+    /// * `q_fn`: queue handler for setting up the queue and its handler
+    /// * `device_fn`: handler called after device is started, run in current
+    ///     context
+    ///
+    /// This one is the preferred interface for creating ublk daemon, and
+    /// is friendly for user, such as, user can customize queue setup and
+    /// io handler, such as setup async/await for handling io command.
+    pub fn run_target<T, Q, W>(&self, tgt_fn: T, q_fn: Q, device_fn: W) -> Result<i32, UblkError>
+    where
+        T: FnOnce(&mut UblkDev) -> Result<i32, UblkError>,
+        Q: FnOnce(u16, &UblkDev) + Send + Sync + Clone + 'static,
+        W: FnOnce(&UblkCtrl) + Send + Sync + 'static,
+    {
+        let dev = &Arc::new(UblkDev::new(self.get_name(), tgt_fn, self)?);
+        let handles = self.create_queue_handlers(dev, q_fn);
+
+        self.start_dev(dev)?;
+
+        device_fn(self);
+
+        for qh in handles {
+            qh.join().unwrap_or_else(|_| {
+                eprintln!("dev-{} join queue thread failed", dev.dev_info.dev_id)
+            });
+        }
+
+        //device may be deleted from another context, so it is normal
+        //to see -ENOENT failure here
+        let _ = self.stop_dev();
+
         Ok(0)
     }
 }
